requirements.txt
flask==2.2.3
python-dotenv==1.0.0
openai==0.27.8
Pillow==9.5.0
pytesseract==0.3.10
requests==2.28.2
torch==2.0.0
transformers==4.27.4
numpy==1.24.3
